+++ packages/core/test/fixtures/test-ui.js
@@ -3,6 +3,3 @@ export async function setup (options) {
 
 export async function render (options, state) {
 }
-
-export async function teardown (options) {
-}