+++ packages/ui/src/vue/store/preferences.js
@@ -5,7 +5,7 @@ const appState = window.UIengine.state
 const properties = {
   hljs: document.getElementById('hljs').getAttribute('data-default'),
   locale: document.documentElement.getAttribute('lang'),
-  currentTheme: (appState.config && appState.config.themes && appState.config.themes[0]),
+  currentTheme: (appState.config && appState.config.ui && appState.config.ui.themes && appState.config.ui.themes[0]),
   navigationCollapsed: false,
   navigationItemsCollapsed: {},
   searchCollapsed: true,